Adjust the sequece of checks in maybeStopDaemon
The method NsdService#maybeStopDaemon is currently being executed
regardless of whether the daemon has been started or not.
However, the mMDnsManager object is always null after Android V.
This means that the null check within the method will always be
triggered, resulting in a wtf log entry. To address this, the
sequence of checks should be adjusted so that the mMDnsManager
null check is performed only after verifying that the daemon has
indeed started up.

Fix: 318874425
Test: Run the NsdManagerTest and check no wtf logs.
Change-Id: Id5bba3a5c6b54df214761ad382f81a6f0a79a907