Log ping packet and reply and use ByteArray.toHexString
ByteArray.toHexString is experimental stdlib API, so it requires a
special annotation to be used. If there is a good reason not to use the
experimental API, an extension function "asHexString()" could be added
to.
Unfortunately, ByteArray does not have a companion object, so it is not
possible to add a fun ByteArray.Companion.fromHexString().

Test: TH
Change-Id: Ifedc998e95d6a1e4cdeb5bd9982b9456e64e649c